[RENAME]style community 폴더의 대소문자 변경 #NADA-49
