+++ src/containers/community/postWrite/Content.js
@@ -1,4 +1,4 @@
-import { ContentStyle } from '../../../styles/mommunity/PostWriteStyle';
+import { ContentStyle } from '../../../styles/community/PostWriteStyle';
 
 export const Content = ({ onChange, inputValue }) => {
   return (
