+++ src/styles/community/BarModalStyle.js
