+++ src/styles/community/CategoryStyle.js
