+++ src/styles/community/CommentBarStyle.js
