+++ src/styles/community/CommunityBarStyle.js
