+++ src/styles/community/CommunityStyle.js
