+++ src/styles/community/NoticeModalStyle.js
