+++ src/styles/community/PostDetailStyle.js
