+++ src/styles/community/PostWriteButtonStyle.js
