+++ src/styles/community/PostWriteStyle.js
