+++ src/styles/community/postItemStyle.js
